fixed suggestions
