+++ src/components/acter/cropper-editor.tsx
@@ -6,7 +6,12 @@ import Cropper from 'react-cropper'
 import 'cropperjs/dist/cropper.css'
 
 const useStyles = makeStyles((theme: Theme) => ({
-  container: {},
+  container: {
+    width: 250,
+    top: 0,
+    left: 0,
+    position: 'absolute',
+  },
   cropBtn: {
     backgroundColor: theme.palette.primary.main,
     color: 'white',
@@ -18,7 +23,7 @@ const useStyles = makeStyles((theme: Theme) => ({
 export interface ImageCropperProps {
   image: any
   aspectRatio: any
-  handleCrop: (cropData: any, canvasData: any, imagefile: any) => void
+  handleCrop: (cropData: any, imagefile: any) => void
 }
 
 export const ImageCropper: FC<ImageCropperProps> = (props) => {
@@ -26,15 +31,45 @@ export const ImageCropper: FC<ImageCropperProps> = (props) => {
   const classes = useStyles()
   const [cropper, setCropper] = useState<any>()
 
-  const handleImageCrop = () =>
-    handleCrop(
-      cropper.getData(),
-      cropper.getCanvasData(),
-      cropper.getCroppedCanvas().toDataURL()
+  const calculateCropData = () => {
+    const cropData = cropper.getData()
+    const canvasData = cropper.getCanvasData()
+    /* Ratio of selected crop area. */
+    const cropAreaRatio = cropData.height / cropData.width
+
+    /* Center point of crop area in percent. */
+    const percentX = (cropData.x + cropData.width / 2) / canvasData.naturalWidth
+    const percentY =
+      (cropData.y + cropData.height / 2) / canvasData.naturalHeight
+
+    /* Calculate available space round image center position. */
+    const cx = percentX > 0.5 ? 1 - percentX : percentX
+    const cy = percentY > 0.5 ? 1 - percentY : percentY
+
+    /* Calculate image rectangle respecting space round image from crop area. */
+    let width = canvasData.naturalWidth
+    let height = width * cropAreaRatio
+    if (height > canvasData.naturalHeight) {
+      height = canvasData.naturalHeight
+      width = height / cropAreaRatio
+    }
+    const rectWidth = cx * 2 * width
+    const rectHeight = cy * 2 * height
+
+    /* Calculate zoom. */
+    const zoom = Math.max(
+      rectWidth / cropData.width,
+      rectHeight / cropData.height
     )
 
+    return { center: { X: percentX, Y: percentY }, cropAreaRatio, zoom }
+  }
+
+  const handleImageCrop = () =>
+    handleCrop(calculateCropData(), cropper.getCroppedCanvas().toDataURL())
+
   return (
-    <div>
+    <div className={classes.container}>
       <Cropper
         src={image !== {} ? image : ''}
         style={{ maxHeight: 200 }}
